fix(frontend): mounted guard and correct user type on home page (#158)
* fix(frontend): add mounted guard and fix user type in page.tsx

- Add `mounted` state flag to prevent auth CTA flash on first render:
  before useEffect fires, the primary CTA renders a size-matched
  transparent placeholder instead of flashing 'Sign In →' for
  authenticated users (matches pattern in dashboard/page.tsx and
  AllModulesButton.tsx)
- Remove `name` property from inline parsedUser type and its
  corresponding fallback in the displayName cascade; domain.User
  has no `name` field — the correct fields are full_name, username,
  and email

* Update frontend/app/page.tsx

* refactor(frontend): centralize stored user parsing and type

---------
