+++ frontend/app/page.tsx
@@ -5,10 +5,13 @@ import Image from 'next/image'
 import { useEffect, useState } from 'react'
 
 export default function Home() {
+  const [mounted, setMounted] = useState(false)
   const [isAuthenticated, setIsAuthenticated] = useState(false)
   const [signedInAs, setSignedInAs] = useState<string | null>(null)
 
   useEffect(() => {
+    setMounted(true)
+
     const rawToken = localStorage.getItem('axiom_token')
     const normalizedToken = rawToken?.replace(/^Bearer\s+/i, '').trim() ?? ''
     const hasValidToken = normalizedToken !== '' && normalizedToken !== 'undefined' && normalizedToken !== 'null'
@@ -28,13 +31,11 @@ export default function Home() {
     try {
       const parsedUser = JSON.parse(rawUser) as {
         full_name?: string
-        name?: string
         username?: string
         email?: string
       }
       const displayName =
         (parsedUser.full_name || '').trim() ||
-        (parsedUser.name || '').trim() ||
         (parsedUser.username || '').trim() ||
         (parsedUser.email || '').trim()
       const email = (parsedUser.email || '').trim()
@@ -88,14 +89,20 @@ export default function Home() {
                 Welcome to Axiom
               </h2>
               <p className="text-gray-700 dark:text-gray-200 mb-6">
-                {isAuthenticated
+                {!mounted
+                  ? 'Loading your experience...'
+                  : isAuthenticated
                   ? 'Welcome back. Continue to your modules or browse public reference data.'
                   : 'Choose where you want to go: sign in for protected features, or browse public reference data.'}
               </p>
             </div>
 
             <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-              {isAuthenticated ? (
+              {!mounted ? (
+                <span className="inline-flex items-center justify-center rounded-md bg-blue-600/40 text-transparent px-6 py-3 font-semibold select-none" aria-hidden="true">
+                  &nbsp;
+                </span>
+              ) : isAuthenticated ? (
                 <Link
                   href="/dashboard"
                   className="inline-flex items-center justify-center rounded-md bg-blue-600 hover:bg-blue-700 text-white px-6 py-3 font-semibold transition-colors"
@@ -118,9 +125,9 @@ export default function Home() {
               </Link>
             </div>
             <div className="mt-6 text-sm text-gray-500 dark:text-gray-400">
-              {isAuthenticated
+              {mounted && (isAuthenticated
                 ? `You are signed in${signedInAs ? ` as ${signedInAs}` : ''}.`
-                : 'Protected modules are available after sign-in.'}
+                : 'Protected modules are available after sign-in.')}
             </div>
           </div>
         </section>
